+++ CHANGELOG.md
@@ -13,6 +13,9 @@ DO NOT ADD CHANGES HERE - ADD THEM USING log_change.sh
 ~~~
 
 
+* Bug **#5665** : Change the basis for time variable replacement from now() to the meta create time when the stream store uses S3. The S3Appender still used now(). Also fix a bug with use of pipeline scoped objects outside of pipeline scope.
+
+
 ## [v7.13-beta.6] - 2026-07-14
 
 * Bug **#5647** : Fix user entered name being ignored when creating a new volume group.